requirements.txt
langchain==0.3.26
flask==3.1.0
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-google-genai==0.1.6
langchain-community==3.0.24
-e .